requirement.txt
torchvision
torch==+cu113  # Replace cu113 with the appropriate CUDA version for your system
streamlit
numpy
matplotlib
pillow